backend/requirements.txt
fastapi==0.119.0
uvicorn==0.22.0
motor==3.6.0
pymongo==4.9.1
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator>=2.0.0
